// src/app/static/src/generated.d.ts
export interface CalibrateDataResponse {
  data: {
    area_id: string;
    sex: string;
    age_group: string;
    calendar_quarter: string;
    indicator: string;
    mode: number | null;
    mean: number | null;
    lower: number | null;
    upper: number | null;
    [k: string]: any;
  }[];
  [k: string]: any;
}
// ...
export interface FilterOption {
  label: string;
  id: string;
  description?: string;
}
// ...
export interface IndicatorMetadata {
  indicator: string;
  value_column: string;
  error_low_column?: string;
  error_high_column?: string;
  indicator_column?: string;
  indicator_value?: string;
  indicator_sort_order?: number;
  name: string;
  min: number;
  max: number;
  colour: string;
  invert_scale: boolean;
  scale: number;
  accuracy: number | null;
  format: string;
}